// Frontend/client-interop/playback-state-changed-event.d.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */
import { IInteropEvent } from "./i-interop-event";
import { PlaybackState } from "./playback-state";
import { InteropEvent } from "./interop-event";
export declare class PlaybackStateChangedEvent implements IInteropEvent {
    playbackState: PlaybackState;
    event: InteropEvent;
}