// Frontend/client-interop/track-changed-event.d.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */
import { IInteropEvent } from "./i-interop-event";
import { InteropEvent } from "./interop-event";
export declare class TrackChangedEvent implements IInteropEvent {
    trackPath: string;
    duration: number;
    event: InteropEvent;
}