// Frontend/client-interop/track-position-changed-event.d.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */
import { IInteropEvent } from "./i-interop-event";
import { InteropEvent } from "./interop-event";
export declare class TrackPositionChangedEvent implements IInteropEvent {
    trackPosition: number;
    visualizationData: number[];
    event: InteropEvent;
}