// Frontend/client-interop/volume-changed-event.d.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */
import { IInteropEvent } from "./i-interop-event";
import { InteropEvent } from "./interop-event";
export declare class VolumeChangedEvent implements IInteropEvent {
    volume: number;
    event: InteropEvent;
}